[Refactor] 가독성 향상을 위한 폴더 구조 정리
