+++ lib/presentation/pages/home/components/current_weather_forecast.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import 'core/theme/constant/app_colors.dart';
+import '../../../../core/theme/constant/app_colors.dart';
 
 class CurrentWeatherForecast extends StatelessWidget {
   const CurrentWeatherForecast({super.key});
